requirements.txt
Django>=5.1
djangorestframework>=3.14.0
drf-spectacular>=0.27.0
Pillow>=10.0.0
gunicorn>=21.2.0
whitenoise>=6.6.0
python-decouple>=3.8
dj-database-url>=2.1.0
